refactor: bot
