+++ src/bot_client/types.ts
@@ -1,0 +1,15 @@
+export type TBotQueries =
+  | 'reg'
+  | 'create_room'
+  | 'create_game'
+  | 'add_ships'
+  | 'start_game'
+  | 'turn'
+  | 'randomAttack'
+  | 'update_winners';
+
+export type TAllBotQuery = {
+  type: TBotQueries;
+  id: number;
+  data: string;
+};
